fix(hydrate): added deprecation message
