+++ src/runtime/composables/useBoosterHydrate.js
@@ -1,5 +1,23 @@
-import boosterHydrate from '#booster/hydrate';
+import { hydrateWhenVisible } from 'vue3-lazy-hydration';
+import { defineAsyncComponent, useRuntimeConfig } from '#imports';
 
 export default function () {
-  return boosterHydrate;
+  return hydrate;
 }
+
+const isDev = process.env.NODE_ENV === 'development';
+
+export const hydrate = component => {
+  const runtimeConfig = useRuntimeConfig();
+
+  return hydrateWhenVisible(wrapComponent(component), {
+    rootMargin: runtimeConfig.public.booster.lazyOffsetComponent || '0%'
+  });
+};
+
+const wrapComponent = component => {
+  if (!(isDev || import.meta.server) && typeof component === 'function') {
+    return defineAsyncComponent(component);
+  }
+  return component;
+};
